1) Don't need base adapter 2) moved adapter's interfaces into separate folder
