--- PocketForzaHorizonCommunity.Back/PocketForzaHorizonCommunity.Back.Database/RepoAdapters/AdapterBase.cs
@@ -1,11 +1,0 @@
-﻿using PocketForzaHorizonCommunity.Back.Database.Entities;
-using PocketForzaHorizonCommunity.Back.Database.Repos;
-
-namespace PocketForzaHorizonCommunity.Back.Database.RepoAdapters;
-
-public abstract class AdapterBase<TEntity> : RepositoryBase<TEntity> where TEntity : EntityBase
-{
-    public AdapterBase(ApplicationDbContext context) : base(context)
-    {
-    }
-}
